Back end

+++ app/Http/Controllers/AdminController.php
@@ -4,10 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Admin;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
-class AdminController extends Controller
+class AdminController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -69,6 +78,8 @@ class AdminController extends Controller
      */
     public function update(Request $request, Admin $admin)
     {
+        Gate::authorize('modify', $admin);
+
         $validator = $request->validate([
             'nom'=> 'required|max:255',
             'prenom'=> 'required|max:255',
@@ -91,8 +102,10 @@ class AdminController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Admin $admin)
+    public function destroy(Admin $admin, Request $request)
     {
+        Gate::authorize('modify', $request);
+
         $admin->delete();
         $data = [
             'status'=> 200,
